poc/pyproject.toml
[project]
name = "chief-of-staff"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.1",
    "fastapi>=0.128.4",
    "openai>=2.17.0",
    "openai-agents>=0.8.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]
